garbage ai/requirements.txt
streamlit
matplotlib
pandas
numpy
# add any other libraries you are using
Pillow
scikit-learn